# src/api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Api JobQuest",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:8800",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/api/upload": {
      "post": {
        "tags": ["Lưu hình"],
        "description": "",     "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/signupEmail": {
      "post": {
        "tags": ["Đăng ký danh sách gửi thư"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
    "/api/company/": {
      "get": {
        "tags": ["Nhà tuyển dụng"],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
...
    "/api/job/": {
      "get": {
        "tags": ["Công việc"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "post": {
        "tags": ["Công việc"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idField": {
                  "example": "any"
                },
                "idProvince": {
                  "example": "any"
                },
                "nameJob": {
                  "example": "any"
                },
                "request": {
                  "example": "any"
                },
                "desc": {
                  "example": "any"
                },
                "other": {
                  "example": "any"
                },
                "salaryMin": {
                  "example": "any"
                },
                "salaryMax": {
                  "example": "any"
                },
                "sex": {
                  "example": "any"
                },
                "typeWork": {
                  "example": "any"
                },
                "education": {
                  "example": "any"
                },
                "experience": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": ["Công việc"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idField": {
                  "example": "any"
                },
                "idProvince": {
                  "example": "any"
                },
                "nameJob": {
                  "example": "any"
                },
                "request": {
                  "example": "any"
                },
                "desc": {
                  "example": "any"
                },
                "other": {
                  "example": "any"
                },
                "salaryMin": {
                  "example": "any"
                },
                "salaryMax": {
                  "example": "any"
                },
                "sex": {
                  "example": "any"
                },
                "typeWork": {
                  "example": "any"
                },
                "education": {
                  "example": "any"
                },
                "experience": {
                  "example": "any"
                },
                "idJob": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Công việc"],
        "description": "",
        "parameters": [
          {
            "name": "idJob",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/api/fields/": {
      "get": {
        "tags": ["Ngành nghề"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/provinces/": {
      "get": {
        "tags": ["Tỉnh"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/follow/": {
      "post": {
        "tags": ["Theo dõi công ty"],
        "description": "",
        "parameters": [
          {
            "name": "idCompany",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Theo dõi công ty"],
        "description": "",
        "parameters": [
          {
            "name": "idCompany",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/save/user/": {
      "get": {
        "tags": ["Lưu công việc"],
        "description": "",
        "parameters": [
          {
            "name": "idJob",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/save/": {
      "get": {
        "tags": ["Lưu công việc"],
        "description": "",
        "parameters": [
          {
            "name": "idUser",
            "in": "query",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "post": {
        "tags": ["Lưu công việc"],
        "description": "",
        "parameters": [
          {
            "name": "idJob",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Lưu công việc"],
        "description": "",
        "parameters": [
          {
            "name": "idJob",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/apply/user": {
      "get": {
        "tags": ["Ứng tuyển"],
        "description": "",
        "parameters": [
          {
            "name": "idJob",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/apply/": {
      "get": {
        "tags": ["Ứng tuyển"],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "post": {
        "tags": ["Ứng tuyển"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "idJob": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "letter": {
                  "example": "any"
                },
                "cv": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}